Split watch add flow views into one file per type
The add flow's helper views lived alongside WatchConfigAddView in a
single file, which conflicts with the one type per file convention. Each
now has its own file and a preview.

The entity flow's area screen is renamed to
WatchConfigAddEntityAreaFilterView: it filters the entity list by area,
which is a different thing from the new list that adds an area as an
item, and both were called WatchConfigAddAreaListView.
